update: 주일 주보 상단바 나오지 않는 버그 수정
