+++ client/src/components/Header/index.tsx
@@ -28,13 +28,6 @@ export default function Header() {
 
   const DrawerItems: Array<DrawerItemsType> = []
 
-  DrawerItems.push({
-    title: "주일 주보",
-    icon: <EventNoteIcon fontSize="small" sx={{ color: "#667eea" }} />,
-    path: "/bulletin/",
-    type: "menu",
-  })
-
   if (isLogin) {
     DrawerItems.push({
       title: "나의 정보 수정",
@@ -55,6 +48,16 @@ export default function Header() {
       } as DrawerItemsType)
     }
   }
+
+  DrawerItems.push({
+    type: "divider",
+  })
+  DrawerItems.push({
+    title: "주일 주보",
+    icon: <EventNoteIcon fontSize="small" sx={{ color: "#667eea" }} />,
+    path: "/common/bulletin/",
+    type: "menu",
+  })
 
   if (authUserData?.role.Leader) {
     DrawerItems.push({
@@ -100,6 +103,10 @@ export default function Header() {
       })
     }
   }
+
+  DrawerItems.push({
+    type: "divider",
+  })
 
   return (
     <Stack
